Move GetSyncFilesFromAttributes  where it is used
